requirements.txt




###################################################################################

# # ===============================
# # FastAPI Framework
# # ===============================
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# # ===============================
# # OpenAI (GPT-4o Vision compatible)
# # ===============================
# openai>=1.47.0

# # ===============================
# # PDF & OCR
# # ===============================
# PyPDF2==3.0.1
# pdf2image==1.16.3
# Pillow==10.1.0

# # ===============================
# # Database
# # ===============================
# pymongo>=4.11.0,<5.0.0
# motor>=3.7.0

# # ===============================
# # AWS
# # ===============================
# boto3==1.29.7

# # ===============================
# # Data Processing
# # ===============================
# numpy==1.26.2
# pandas==2.1.3

# # ===============================
# # Visualization
# # ===============================
# matplotlib==3.8.2
# seaborn==0.13.0

# # ===============================
# # Environment & Config
# # ===============================
# python-dotenv==1.0.0
# pydantic==2.5.0
# pydantic-settings==2.1.0
# python-dateutil==2.8.2

# # ===============================
# # Logging
# # ===============================
# structlog==23.2.0

# # ===============================
# # Testing
# # ===============================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2

# ##################################################################
# # ===============================
# # FastAPI Framework
# # ===============================
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# # ===============================
# # OpenAI (GPT-4o Vision compatible)
# # ===============================
# openai>=1.47.0

# # ===============================
# # PDF & Image OCR
# # ===============================
# PyPDF2==3.0.1
# pdf2image==1.16.3
# Pillow==10.1.0

# # ===============================
# # Database
# # ===============================
# pymongo>=4.11.0,<5.0.0
# motor>=3.7.0

# # ===============================
# # AWS
# # ===============================
# boto3==1.29.7

# # ===============================
# # Data Processing
# # ===============================
# numpy==1.26.2
# pandas==2.1.3

# # ===============================
# # Visualization
# # ===============================
# matplotlib==3.8.2
# seaborn==0.13.0

# # ===============================
# # PHREEQC Python wrapper
# # ===============================
# phreeqpython>=1.5.7
# phreeqpy>=0.6.0

# # ===============================
# # Environment & Config
# # ===============================
# python-dotenv==1.0.0
# pydantic==2.5.0
# pydantic-settings==2.1.0
# python-dateutil==2.8.2

# # ===============================
# # Logging
# # ===============================
# structlog==23.2.0

# # ===============================
# # Testing
# # ===============================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2






# ===============================
# FastAPI Framework
# ===============================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===============================
# OpenAI (GPT-4o Vision compatible)
# ===============================
openai>=1.47.0

# ===============================
# PDF & Image OCR
# ===============================
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0

# ===============================
# Database
# ===============================
pymongo>=4.11.0,<5.0.0
motor>=3.7.0

# ===============================
# AWS
# ===============================
boto3==1.29.7

# ===============================
# Data Processing
# ===============================
numpy==1.26.2
pandas==2.1.3

# ===============================
# Visualization
# ===============================
matplotlib==3.8.2
seaborn==0.13.0

# ===============================
# ❌ REMOVED - CLIENT REQUIREMENT
# phreeqpython - NOT NEEDED (using core PHREEQC via subprocess)
# phreeqpy - NOT NEEDED
# ===============================
# We are using CORE PHREEQC calculation engine directly
# No Python wrapper required

# ===============================
# Environment & Config
# ===============================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# ===============================
# Logging
# ===============================
structlog==23.2.0

# ===============================
# Testing
# ===============================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2